--- backend/obelix-quarry-api/src/main/java/ch/bbw/obelix/quarry/api/repository/MenhirRepository.java
@@ -1,17 +1,0 @@
-package ch.bbw.obelix.quarry.api.repository;
-
-import java.util.List;
-import java.util.UUID;
-
-import ch.bbw.obelix.quarry.api.entity.MenhirEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface MenhirRepository extends JpaRepository<MenhirEntity, UUID> {
-
-	List<MenhirEntity> findByStoneTypeContainingIgnoreCase(String stoneType);
-
-	List<MenhirEntity> findMenhirByDecorativeness(MenhirEntity.Decorativeness decorativeness);
-
-}
